tree.h API: remove support for starting at prefix != ""
Every caller of the read_tree_recursive() function hardcoded a
starting point of "" in the tree. So let's simply remove that
parameter.

The last function to call read_tree_recursive() with a non-"" path was
read_tree_recursive() itself, but that was changed in
ffd31f661d5 (Reimplement read_tree_recursive() using
tree_entry_interesting(), 2011-03-25).

If in the future we need to support recursively reading trees without
starting at the root we can easily add a read_tree_recursive_subdir(),
and make that function a thin wrapper for read_tree_1().

In the meantime there's no reason to keep around what amounts to dead
code, just in case we need it in the future.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>